services/media_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
supabase
boto3
Pillow
python-multipart
